+++ src-docs/src/views/date_picker/super_date_picker.js
@@ -1,0 +1,64 @@
+
+import React, { Component } from 'react';
+
+import {
+  EuiSuperDatePicker,
+} from '../../../../src/components';
+
+export default class extends Component {
+
+  state = {
+    recentlyUsedRanges: [],
+    isLoading: false,
+  }
+
+  onTimeChange = ({ start, end }) => {
+    this.setState((prevState) => {
+      const recentlyUsedRanges = prevState.recentlyUsedRanges.filter(recentlyUsedRange => {
+        const isDuplicate = recentlyUsedRange.start === start && recentlyUsedRange.end === end;
+        return !isDuplicate;
+      });
+      recentlyUsedRanges.unshift({ start, end });
+      return {
+        start,
+        end,
+        recentlyUsedRanges: recentlyUsedRanges.length > 10 ? recentlyUsedRanges.slice(0, 9) : recentlyUsedRanges,
+        isLoading: true,
+      };
+    }, this.startLoading);
+  }
+
+  startLoading = () => {
+    setTimeout(
+      this.stopLoading,
+      1000);
+  }
+
+  stopLoading = () => {
+    this.setState({ isLoading: false });
+  }
+
+  onRefreshChange = ({ isPaused, refreshInterval }) => {
+    this.setState((prevState) => {
+      return {
+        isPaused: isPaused == null ? prevState.isPaused : isPaused,
+        refreshInterval: refreshInterval == null ? prevState.refreshInterval : refreshInterval,
+      };
+    });
+  }
+
+  render() {
+    return (
+      <EuiSuperDatePicker
+        isLoading={this.state.isLoading}
+        start={this.state.start}
+        end={this.state.end}
+        onTimeChange={this.onTimeChange}
+        isPaused={this.state.isPaused}
+        refreshInterval={this.state.refreshInterval}
+        onRefreshChange={this.onRefreshChange}
+        recentlyUsedRanges={this.state.recentlyUsedRanges}
+      />
+    );
+  }
+}