+++ src/components/date_picker/index.js
@@ -5,3 +5,7 @@ export {
 export {
   EuiDatePickerRange,
 } from './date_picker_range';
+
+export {
+  EuiSuperDatePicker,
+} from './super_date_picker';