+++ src/components/date_picker/super_date_picker/index.js
@@ -1,0 +1,5 @@
+import {
+  WrappedEuiSuperDatePicker,
+} from './super_date_picker';
+
+export { WrappedEuiSuperDatePicker as EuiSuperDatePicker };