+++ src/components/date_picker/super_date_picker/super_date_picker.test.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import {
+  EuiSuperDatePicker,
+} from './super_date_picker';
+
+const noop = () => {};
+
+describe('EuiSuperDatePicker', () => {
+  test('is rendered', () => {
+    const component = shallow(
+      <EuiSuperDatePicker
+        onTimeChange={noop}
+      />
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+
+  test('isLoading', () => {
+    const component = shallow(
+      <EuiSuperDatePicker
+        isLoading
+        onTimeChange={noop}
+      />
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+});